baseline to begin refactoring OClassSet as child of OClassImpl and take OClassVersion out of that class hierarchy

+++ binary/src/main/java/com/orientechnologies/binary/runnabletest/TestOdbBinMain.java
@@ -9,11 +9,15 @@ import java.util.Date;
 import java.util.List;
 
 import com.orientechnologies.binary.BinaryDocumentSerializer;
+import com.orientechnologies.binary.FieldSerializeStrategy;
 import com.orientechnologies.binary.OClassSet;
 import com.orientechnologies.binary.OBinaryDocument;
 import com.orientechnologies.binary.OBinProperty;
 import com.orientechnologies.binary.OClassVersion;
+import com.orientechnologies.binary.jackson.JsonSerializer;
 import com.orientechnologies.orient.core.db.document.ODatabaseDocumentTx;
+import com.orientechnologies.orient.core.id.ORID;
+import com.orientechnologies.orient.core.id.ORecordId;
 import com.orientechnologies.orient.core.metadata.schema.OType;
 import com.orientechnologies.orient.core.record.impl.ODocument;
 import com.orientechnologies.orient.core.serialization.serializer.record.ORecordSerializerFactory;
@@ -23,43 +27,64 @@ import com.orientechnologies.orient.server.OServerMain;
 /**
  * 
  * @author Steve Coughlan
- *
+ * 
  */
 public class TestOdbBinMain {
 
 	public static void main(String[] args) throws Exception {
 
+		OBinaryDocument doc;
+
 		ORecordSerializerFactory.instance().register(BinaryDocumentSerializer.NAME, new BinaryDocumentSerializer());
+		FieldSerializeStrategy.setStrategy(JsonSerializer.JSON_STRATEGY);
+
+		boolean isNew;
+		isNew = true;
+		//isNew = false;
 
 		File dir = new File("/home/git/junk/rubbish/orientdb");
-		if (dir.exists())
-			for (File file : dir.listFiles())
-				if (file.isFile())
-					file.delete();
+		ODatabaseDocumentTx db;
 
-		ODatabaseDocumentTx db = new ODatabaseDocumentTx("plocal:" + dir.getAbsolutePath()).create();// .open("admin",
-																										// "admin");
+		if (isNew) {
+			if (dir.exists())
+				for (File file : dir.listFiles())
+					if (file.isFile())
+						file.delete();
+
+			db = new ODatabaseDocumentTx("plocal:" + dir.getAbsolutePath()).create();
+			
+		} else {
+			
+			db = new ODatabaseDocumentTx("plocal:" + dir.getAbsolutePath()).open("admin", "admin");
+			db.getMetadata().getSchema().getClass("employee");
+			ODocument normalDoc;
+			normalDoc = db.load(new ORecordId("#9:0"));
+			System.out.println(normalDoc);
+		}
+
+		
 
 		List<OBinProperty> fields = new ArrayList();
 		fields.add(new OBinProperty());
 
-		OClassSet schemaSet = OClassSet.newSchemaSet("employee");
-		OClassVersion schema = schemaSet.currentSchema();
-		schema.addProperty(new OBinProperty(schema, "fullname", OType.STRING));
-		schema.addProperty(new OBinProperty(schema, "age", OType.INTEGER));
-		schema.addProperty(new OBinProperty(schema, "dateofbirth", OType.DATE));
-		schema.addProperty(new OBinProperty(schema, "nullString", OType.STRING));
-		schema.addProperty(new OBinProperty(schema, "nullFloat", OType.FLOAT));
-		schema.makeImmutable();
+		// OClassSet schemaSet = OClassSet.newSchemaSet("employee");
+		OClassSet schemaSet = (OClassSet) db.getMetadata().getSchema().createClass("employee");
+		OClassVersion clazz = schemaSet.currentSchema();
+		clazz.addProperty(new OBinProperty(clazz, "fullname", OType.STRING));
+		clazz.addProperty(new OBinProperty(clazz, "age", OType.INTEGER));
+		clazz.addProperty(new OBinProperty(clazz, "dateofbirth", OType.DATE));
+		clazz.addProperty(new OBinProperty(clazz, "nullString", OType.STRING));
+		clazz.addProperty(new OBinProperty(clazz, "nullFloat", OType.FLOAT));
+		clazz.makeImmutable();
 
-		db.getMetadata().getSchema().createClass(schema.getName());
-		
-		OBinaryDocument doc = new OBinaryDocument(schema);
-		
-		//OBinaryDocument doc = new OBinaryDocument();
-		//ODocument doc = new ODocument();
+		// db.getMetadata().getSchema().createClass(schema.getName());
+
+		doc = new OBinaryDocument(clazz);
+
+		// OBinaryDocument doc = new OBinaryDocument();
+		// ODocument doc = new ODocument();
 		long _25Years = 25 * 1000l * 60 * 60 * 24 * 365;
-		
+
 		doc.field("fullname", "Luca Garulli", OType.STRING);
 		doc.field("age", 25, OType.INTEGER);
 		doc.field("dateofbirth", new Date(System.currentTimeMillis() - _25Years), OType.DATE);
@@ -69,35 +94,33 @@ public class TestOdbBinMain {
 		doc.field("randomInt", 99, OType.SHORT);
 		doc.field("randomString", "My Random String", OType.STRING);
 
-		System.out.println("old: " + prettyDoc(doc));	
-		
+		System.out.println("old: " + prettyDoc(doc));
+
 		byte[] bytes = new byte[100];
 		BinaryDocumentSerializer ser = new BinaryDocumentSerializer();
-		
+
 		bytes = ser.toStream(doc, false);
 		saveBytes(bytes, new File(dir, "data/serialized.bin"));
 		OBinaryDocument newDoc = (OBinaryDocument) ser.fromStream(bytes, new OBinaryDocument());
-		
-		System.out.println("new: " + prettyDoc(newDoc));	
-		
-		OBinaryDocument partDoc1 = (OBinaryDocument) ser.fromStream(bytes, new OBinaryDocument(), 
-				new String[] {"fullname", "randomInt", "age"});
-		
-		System.out.println("part1: " + prettyDoc(partDoc1));	
-		
-		OBinaryDocument partDoc2 = (OBinaryDocument) ser.fromStream(bytes, new OBinaryDocument(), 
-				new String[] {"dateofbirth", "randomString", "nullFloat", "nullString", });
-		
-		System.out.println("part2: " + prettyDoc(partDoc2));	
-		
-		
-		//db.save(doc, schema.getName());
 
-		
-		//doc = db.load(doc.getIdentity());
-		
-		//System.out.println(doc);		
-		
+		System.out.println("new: " + prettyDoc(newDoc));
+
+		OBinaryDocument partDoc1 = (OBinaryDocument) ser.fromStream(bytes, new OBinaryDocument(), new String[] {
+				"fullname", "randomInt", "age" });
+
+		System.out.println("part1: " + prettyDoc(partDoc1));
+
+		OBinaryDocument partDoc2 = (OBinaryDocument) ser.fromStream(bytes, new OBinaryDocument(), new String[] {
+				"dateofbirth", "randomString", "nullFloat", "nullString", });
+
+		System.out.println("part2: " + prettyDoc(partDoc2));
+
+		db.save(doc, clazz.getName());
+
+		doc = db.load(doc.getIdentity());
+
+		System.out.println(doc);
+
 		db.close();
 
 	}
@@ -108,7 +131,7 @@ public class TestOdbBinMain {
 		StringBuilder sb = new StringBuilder();
 		sb.append(doc.getClassName()).append("{");
 		boolean first = true;
-		for (String field: fields) {
+		for (String field : fields) {
 			if (!first)
 				sb.append(", ");
 			first = false;
@@ -118,14 +141,14 @@ public class TestOdbBinMain {
 		sb.append("}");
 		return sb.toString();
 	}
-	
+
 	private static void saveBytes(byte[] bytes, File file) throws Exception {
 		file.getParentFile().mkdirs();
 		FileOutputStream out = new FileOutputStream(file, false);
 		out.write(bytes);
 		out.close();
 	}
-	
+
 	private static void startServer() throws Exception {
 		OServer server = OServerMain.create();
 		server.startup("<?xml version=\"1.0\" encoding=\"UTF-8\" standalone=\"yes\"?>"
